+++ x-pack/solutions/observability/plugins/observability/public/pages/alert_details/hooks/use_discover_url/get_rule_data/es_query_rule.ts
